pyproject.toml
[project]
name = "task-management-app"
version = "0.1.0"
description = ""
authors = [
    {name = "anatoly.rozhkov1998@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "django-silk (>=5.3.2,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "black (>=24.10.0,<25.0.0)",
    "drf-extended-viewset (>=0.1.1,<0.2.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-livereload-server (>=0.5.1,<0.6.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-jazzmin (>=3.0.1,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-filter (>=25.1,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.co re.masonry.api"

[tool.isort]
indent=4
multi_line_output=3
include_trailing_comma=true
force_grid_wrap = 0
line_length=120
use_parentheses = true
ensure_newline_before_comments = true
known_django='django'
known_rest_framework='rest_framework'
known_api='api'
sections='FUTURE,STDLIB,DJANGO,REST_FRAMEWORK,THIRDPARTY,FIRSTPARTY,LOCALFOLDER,API'
skip_glob='**/migrations/**'